Support building on Mac OS X 10.12 and later in Xcode 8.x. No functional changes.

+++ Classes/OpticalFlowMotionAnalyzer.mm
@@ -187,7 +187,7 @@ static const double MinimumMovementMagnitude = 0.5;
                                        prevPyr,
                                        &*featuresCur.begin(),
                                        featuresPrev,
-                                       featuresCur.size(),
+                                       (int)featuresCur.size(),
                                        cvSize(15, 15),      // pyramid window size
                                        5,                   // number of pyramid levels
                                        featuresPrevFound,
